requirements.txt
beautifulsoup4
click
pyyaml
requests
